phylo: add hardcoded threads for tree and align
Previously used `--nthread auto` which would let each tree/align
command use _all_ available cores. Since this is not done with
Snakemake threads, Snakemake is not aware of this usage and will end up
scheduling of multiple jobs that would oversubscribe. This results in
out of memory errors when running the workflow locally and likely also
the reason for the failed GH Action workflow.¹

This commit adds a hardcoded Snakemake thread so that Snakemake can
properly schedule the jobs based on available cores. If the single
threaded jobs take too long, then we can bump the threads and/or
transition to running phylo on AWS Batch.

¹ <https://github.com/nextstrain/chikv/actions/runs/20007579231>